requirements.txt
asn1crypto==1.5.1
boto3==1.38.45
botocore==1.38.45
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
filelock==3.18.0
greenlet==3.2.3
idna==3.10
jmespath==1.0.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
